Move DatePicker component to be under the 'internal' folder
