+++ src/components/ButtonDatePicker.tsx
@@ -1,7 +1,7 @@
 import { useRef } from "react";
 import { useMenuTrigger } from "react-aria";
 import { useMenuTriggerState } from "react-stately";
-import { DatePickerProps } from "src/components/DatePicker";
+import { DatePickerProps } from "src/components/internal/DatePicker";
 import { DatePickerOverlay } from "src/components/internal/DatePickerOverlay";
 import { isTextButton, OverlayTrigger, OverlayTriggerProps } from "src/components/internal/OverlayTrigger";
 import { useTestIds } from "src/utils";
